sue: enable booting by loading over SPI
This patch enables creation of an SOF image, that can be loaded over
SPI and run on Sue Creek platforms. With this patch an image still has
to be manually created by padding the boot loader and concatenating it
with the SOF proper rimage (see comments in memory.h). It also contains
a hard-coded 9KiB offset to the beginning of the manifest. More work is
required to fully automate image creation.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>