cavs: eliminate per-core interrupt controller descriptors
Currently cascading interrupt controller descriptors are registered
per core. This makes it possible to register multiple interrupt
handlers for the same IRQ to be run on different cores. However, the
same is possible with a single controller descriptor per interrupt.
In this case care should be taken to keep shared data synchronised
across cores.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>