interrupt: make enable counters per core
On cAVS platforms we deal with two kinds of interrupt controllers:
generic ones, which have a certain number of interrupt input lines,
each of which can be masked individually, and a single interrupt
output, and specialised so called "level" interrupt controllers,
which can mask each input line separately for each DSP core. To
efficiently handle the latter ones we have to use per-core interrupt
enable counters on both the input and the output sides. This patch
adds such handling and a differentiation between the two controller
types.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>