llext: move the llext pointer to struct lib_manager_mod_ctx
The LLEXT context is currently bound to the Module Adapter API which
shouldn't be the case because we also need generic LLEXT-based
loadable SOF objects. This patch moves the context to struct
lib_manager_mod_ctx which is safe, because the library manager is
always needed when dynamically loading modules.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>